+++ Cargo.toml
@@ -39,6 +39,11 @@ path = "tests/blockchain.rs"
 required-features = ["random"]
 
 [[test]]
+name = "test-contract"
+path = "tests/contract.rs"
+required-features = ["random"]
+
+[[test]]
 name = "test-debug"
 path = "tests/debug.rs"
 required-features = ["debug"]
