+++ src/interpreter/executors/instruction.rs
@@ -322,6 +322,11 @@ where
                 self.store_word(a, b, imm)?;
             }
 
+            OpcodeRepr::BAL => {
+                self.gas_charge(GAS_BAL)?;
+                self.contract_balance(ra, b, c)?;
+            }
+
             OpcodeRepr::BHEI => {
                 self.gas_charge(GAS_BHEI)?;
                 self.alu_set(ra, self.block_height() as Word)?;
